Backend (Python)/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = ""
authors = ["Mohsin <msnsh12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
passlib = "^1.7.4"
google-generativeai = "^0.8.1"
ipython = "^8.27.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.18"
psycopg = {extras = ["binary"], version = "^3.1.19"}
aiokafka = "^0.10.0"
protobuf = "^5.26.1"